Move imgsift_client function
